requirements.txt
accelerate==1.0.1
backoff==2.2.1
datasets==3.0.1
evaluate==0.4.3
fastapi==0.115.12
gradio==5.27.1
matplotlib==3.10.1
numpy==2.2.5
openai==1.76.2
pandas==2.2.3
peft==0.15.2
Pillow==11.2.1
POT==0.9.5
PyYAML==6.0.2
qdrant_client==1.14.2
Requests==2.32.3
scikit_learn==1.6.1
sentence_transformers==4.1.0
sentencex==0.6.1
setuptools==75.8.0
shortuuid==1.0.13
tiktoken==0.9.0
sentencepiece==0.2.0
torch==2.3.1
tqdm==4.67.1
transformers==4.45.2
uvicorn==0.34.2
wandb==0.19.10
